+++ CHANGELOG.md
@@ -67,3 +67,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Config Loader
 - Human Readable to bridgehead
 - Show minimal information in project form
+- Show User TODOs